requirements.txt
Flask==3.0.0
python-pptx==0.6.23
python-docx==1.1.0
openai==1.12.0
requests==2.31.0
Werkzeug==3.0.1
Pillow==10.2.0
pytest==8.0.0
python-dotenv==1.0.0